CareCompassApp/backend/requirements.txt
flask
flask-cors
firebase-admin
cryptography
google-generativeai
pyyaml
